requirements.txt
Flask>=2.0.0
pywebview>=4.0
pyobjc-framework-Cocoa
pyobjc-framework-WebKit
py2app
